Refactor: Seeding functions
