+++ backend/src/db/seeds/seeders/index.js
@@ -1,0 +1,7 @@
+import seedRoles from "./seedRoles";
+import seedTasks from "./seedTasks";
+import seedTaskUsers from "./seedTaskUsers";
+import seedUnits from "./seedUnits";
+import seedUsers from "./seedUsers";
+
+export { seedRoles, seedTasks, seedTaskUsers, seedUnits, seedUsers };
